feat: 🎸 add design
