+++ src/components/SwitchLang.astro
@@ -23,7 +23,7 @@ const currentLang = Astro.params.lang as TLanguage;
   </label>
 
   <ul
-    class="absolute top-[calc(100%+8px)] left-0 w-max px-3 py-2 bg-dark border border-white rounded-md shadow-lg space-y-1 opacity-0 scale-95 peer-checked:opacity-100 peer-checked:scale-100 transition-all duration-200 origin-top z-50"
+    class="absolute top-[calc(100%+8px)] left-0 w-max px-3 py-2 bg-dark border border-white rounded-md shadow-lg space-y-1 opacity-0 scale-95 pointer-events-none peer-checked:opacity-100 peer-checked:scale-100 peer-checked:pointer-events-auto transition-all duration-200 origin-top z-50"
   >
     {
       Object.entries(languages).map(([code, value]) => (
