+++ src/components/SwitchThemes.astro
@@ -2,8 +2,8 @@
 import { ChevronDown } from "@lucide/astro";
 
 const themes = [
-  { id: "theme1", from: "#00F5A0", to: "#00D9FF" },
-  { id: "theme2", from: "#0066FF", to: "#00FFC6" },
+  { id: "theme1", from: "#0066FF", to: "#00FFC6" },
+  { id: "theme2", from: "#00F5A0", to: "#00D9FF" },
   { id: "theme3", from: "#00FFE0", to: "#FFD500" },
 ];
 ---
@@ -15,14 +15,14 @@ const themes = [
     class="cursor-pointer flex items-center justify-center gap-2 border border-white rounded-lg p-2"
   >
     <div
-      class="size-10 bg-gradient-to-br from-primary to-secondary rounded-full"
+      class="size-8 bg-gradient-to-br from-primary to-secondary rounded-full"
     >
     </div>
     <ChevronDown class="text-white" size={18} />
   </label>
 
   <div
-    class="absolute top-[calc(100%+8px)] left-1 w-fit p-2 border border-white bg-dark rounded-lg shadow-lg opacity-0 scale-95 peer-checked:opacity-100 peer-checked:scale-100 transition-all duration-300 origin-top z-50"
+    class="absolute top-[calc(100%+8px)] left-1 w-fit p-2 border border-white bg-dark rounded-lg shadow-lg opacity-0 scale-95 pointer-events-none peer-checked:opacity-100 peer-checked:scale-100 peer-checked:pointer-events-auto transition-all duration-300 origin-top z-50"
   >
     <div class="flex flex-col items-center gap-2">
       {
@@ -57,7 +57,7 @@ const themes = [
   if (savedTheme) {
     document.documentElement.setAttribute("data-theme", savedTheme);
   } else {
-    localStorage.setItem("data-theme", "theme1");
+    localStorage.setItem("theme", "theme1");
   }
 
   window.applyTheme = (theme) => {
