+++ src/components/background/points-left.astro
@@ -3,55 +3,3 @@
   aria-hidden="true"
 >
 </div>
-
-<!-- <div class="fondo-puntos"></div>
-
-<style>
-  .fondo-puntos {
-    position: relative;
-    width: 100%;
-    height: 100vh;
-    background-color: #0a0a0a;
-    overflow: hidden;
-  }
-  /* ANIMACIÓN */
-  @keyframes pulso {
-    0%,
-    100% {
-      opacity: 0.15;
-      transform: scale(1);
-    }
-    50% {
-      opacity: 0.3;
-      transform: scale(1.02);
-    }
-  }
-  /* Triángulo de puntos verdes a la izquierda */
-  .fondo-puntos::before {
-    content: "";
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 50%;
-    height: 100%;
-    background-image: radial-gradient(circle, #00ffcc 1.2px, transparent 1.2px);
-    background-size: 14px 14px;
-    clip-path: polygon(0 0, 100% 50%, 0 100%);
-    animation: pulso 6s ease-in-out infinite;
-    transform-origin: left center;
-  }
-  /* Triángulo de puntos azules a la derecha */
-  .fondo-puntos::after {
-    content: "";
-    position: absolute;
-    top: 0;
-    right: 0;
-    width: 50%;
-    height: 100%;
-    background-image: radial-gradient(circle, #0099ff 1.2px, transparent 1.2px);
-    background-size: 14px 14px;
-    clip-path: polygon(100% 0, 0 50%, 100% 100%);
-    animation: pulso 6s ease-in-out infinite;
-    transform-origin: right center;
-  }
-</style> -->
