+++ src/components/background/points-right.astro
@@ -1,5 +1,6 @@
 <div
-  class="absolute right-0 top-1/2 -translate-y-1/2 h-[200%] max-h-[200vw] w-[100vw] bg-primary mask mask-[url('/points-right.png')] mask-right mask-contain mask-no-repeat"
+  class="-z-20 absolute right-0 top-1/2 -translate-y-1/2 h-[200%] max-h-[200vw] w-[100vw] bg-primary mask mask-[url('/points-right.png')] mask-right mask-contain mask-no-repeat overflow-hidden"
   aria-hidden="true"
+  id="bg-with-sparkles"
 >
 </div>
