+++ src/i18n/fr.ts
@@ -36,29 +36,29 @@ export default {
         title: "Que fait l'IA ?",
         cards: [
           {
-            image: "/icons/code-review.svg",
+            image: "/icons/code-review.png",
             label: "Revue de code automatique",
             description:
               "Améliorez votre code selon des standards comme SonarQube",
           },
           {
-            image: "/icons/intelligent-comments.svg",
+            image: "/icons/intelligent-comments.png",
             label: "Commentaires intelligents sur les PR",
             description: "Édition collaborative avec l'IA",
           },
           {
-            image: "/icons/contextual-chat.svg",
+            image: "/icons/contextual-chat.png",
             label: "Chat contextuel",
             description:
               "Une IA qui comprend votre code et répond à vos questions",
           },
           {
-            image: "/icons/documentation.svg",
+            image: "/icons/documentation.png",
             label: "Documentation technique automatique",
             description: "Toujours à jour, sans rédaction manuelle",
           },
           {
-            image: "/icons/github.svg",
+            image: "/icons/github.png",
             label: "Intégration sécurisée avec GitHub",
             description: "OAuth et dépôts privés protégés",
           },
@@ -107,22 +107,22 @@ export default {
         title: "À qui s'adresse cette solution ?",
         card: [
           {
-            img: "/icons/dev-team.svg",
+            img: "/icons/dev-team.png",
             label: "Développeurs et équipes logicielles",
             isGradient: false,
           },
           {
-            img: "/icons/startups.svg",
+            img: "/icons/startups.png",
             label: "Startups recherchant la rapidité sans sacrifier la qualité",
             isGradient: true,
           },
           {
-            img: "/icons/consultant.svg",
+            img: "/icons/consultant.png",
             label: "Cabinets de conseil qui doivent maintenir des standards",
             isGradient: true,
           },
           {
-            img: "/icons/education.svg",
+            img: "/icons/education.png",
             label: "Plateformes éducatives enseignant le code propre",
             isGradient: false,
           },
