+++ src/modules/home/benefits/benefits.astro
@@ -2,7 +2,7 @@
 import type { TLanguage } from "@i18n/config/interfaces/languaje.interface";
 import { translator } from "@i18n/config/translator";
 import PointsLeft from "@components/background/points-left.astro";
-import { BenefitsItem } from "./benefit-item";
+import BenefitsItem from "./benefit-item.astro";
 
 const lang = Astro.params.lang as TLanguage;
 const t = translator(lang);
@@ -18,21 +18,21 @@ const content = t.pages.home.benefits;
       class="font-grandis heading-h3 text-center md:text-left [&>span]:mb-4 [&>span]:block mb-4"
       set:html={content.title}
     />
-    <ul
-      class="relative px-8 py-16 w-full md:w-auto flex flex-col gap-8 rounded-lg overflow-hidden"
-    >
+    <div class="relative w-full md:w-auto">
       <div
         class="-z-10 absolute w-full h-full left-0 top-0 bg-radial from-white from-30% to-gray-500 opacity-15"
       >
       </div>
-      {
-        content.card.map((i) => (
-          <BenefitsItem percent={i.percent} client:visible>
-            {i.text}
-          </BenefitsItem>
-        ))
-      }
-    </ul>
+      <ul
+        class="px-8 py-16 w-full flex flex-col gap-8 rounded-lg overflow-hidden"
+      >
+        {
+          content.card.map((i) => (
+            <BenefitsItem percent={i.percent} label={i.text} />
+          ))
+        }
+      </ul>
+    </div>
   </div>
   <PointsLeft />
 </section>
