+++ src/modules/home/hero/hero.astro
@@ -1,11 +1,9 @@
 ---
-import Logo from "@components/Logo.astro";
-import SwitchLang from "@components/SwitchLang.astro";
-import SwitchThemes from "@components/SwitchThemes.astro";
 import type { TLanguage } from "@i18n/config/interfaces/languaje.interface";
 import { translator } from "@i18n/config/translator";
-import Waves from "./waves.astro";
 import ChatBot from "./chat-bot/chat-bot";
+import Navbar from "./navbar.astro";
+import Waves from "./waves.astro";
 
 const lang = Astro.params.lang as TLanguage;
 const t = translator(lang);
@@ -16,13 +14,7 @@ const content = t.pages.home.hero;
   class="z-10 relative container mx-auto py-4 h-[80vh] max-h-[900px] rounded-b-3xl flex flex-col"
 >
   <Waves />
-  <nav class="flex items-center justify-between">
-    <Logo />
-    <div class="flex items-center gap-4">
-      <SwitchLang languages={t.languageOptions} />
-      <SwitchThemes />
-    </div>
-  </nav>
+  <Navbar />
   <div class="grow flex items-center justify-center">
     <div class="max-w-[70%]">
       <h1 class="text-white text-center heading-h2">
