+++ src/modules/home/lets-get-started/lets-get-started.astro
@@ -32,7 +32,7 @@ const content = t.pages.home.letsGetStarted;
       {content.description}
     </p>
     <button
-      class="bg-primary rounded-[36px] px-16 py-2 font-bold mt-16 text-custom-lg cursor-pointer hover:shadow-xl hover:shadow-primary duration-[1s]"
+      class="bg-primary rounded-[36px] px-16 py-2 font-bold mt-16 text-custom-lg text-dark cursor-pointer hover:shadow-xl hover:shadow-primary duration-[1s]"
       onclick={`document.getElementById('contactUs')?.scrollIntoView({ behavior: 'smooth' })`}
       aria-label="Ir a un formulario de contacto para empezar ahora"
     >
