+++ src/modules/home/what-are-doing/carrusel.astro
@@ -18,12 +18,21 @@ const carruselItems = content.cards;
   >
     <ChevronLeft />
   </button>
+
   <div class="relative md:max-w-xl overflow-hidden">
     <div
       id="carousel-track"
       class="flex transition-transform duration-500 ease-in-out w-[300px] md:w-auto mx-auto"
-      style="transform: translateX(0%)"
+      style="transform: translateX(-100%)"
     >
+      <div class="w-full shrink-0">
+        <CarruselCard
+          image={carruselItems[carruselItems.length - 1].image}
+          label={carruselItems[carruselItems.length - 1].label}
+          description={carruselItems[carruselItems.length - 1].description}
+        />
+      </div>
+
       {
         carruselItems.map((e) => (
           <div class="w-full shrink-0">
@@ -35,8 +44,17 @@ const carruselItems = content.cards;
           </div>
         ))
       }
+
+      <div class="w-full shrink-0">
+        <CarruselCard
+          image={carruselItems[0].image}
+          label={carruselItems[0].label}
+          description={carruselItems[0].description}
+        />
+      </div>
     </div>
   </div>
+
   <button
     class="z-10 hidden md:block pointer-events-auto p-2 rounded-full border border-white text-white bg-white/10 hover:bg-black/70 transition cursor-pointer"
     onclick="nextSlide()"
@@ -45,8 +63,7 @@ const carruselItems = content.cards;
     <ChevronRight />
   </button>
 
-  <!-- buttons movile -->
-  <div class="md:hidden flex justify-between">
+  <div class="md:hidden flex justify-between mt-4">
     <button
       class="z-10 pointer-events-auto p-2 rounded-full border border-white text-white bg-white/10 hover:bg-black/70 transition mr-4 cursor-pointer"
       onclick="prevSlide()"
@@ -67,24 +84,41 @@ const carruselItems = content.cards;
 <script is:inline>
   const track = document.getElementById("carousel-track");
   const totalSlides = track.children.length;
-  let current = 0;
+  let current = 1;
 
-  function updateCarousel() {
+  function updateCarousel(smooth = true) {
+    if (!smooth) {
+      track.style.transition = "none";
+    } else {
+      track.style.transition = "transform 0.5s ease-in-out";
+    }
     const percent = -(100 * current);
     track.style.transform = `translateX(${percent}%)`;
   }
 
   function nextSlide() {
-    if (current < totalSlides - 1) {
-      current++;
-      updateCarousel();
-    }
+    if (current >= totalSlides - 1) return;
+    current++;
+    updateCarousel();
   }
 
   function prevSlide() {
-    if (current > 0) {
-      current--;
-      updateCarousel();
-    }
+    if (current <= 0) return;
+    current--;
+    updateCarousel();
   }
+
+  track.addEventListener("transitionend", () => {
+    if (current === totalSlides - 1) {
+      current = 1;
+      updateCarousel(false);
+    } else if (current === 0) {
+      current = totalSlides - 2;
+      updateCarousel(false);
+    }
+  });
+
+  setInterval(() => {
+    nextSlide();
+  }, 5000);
 </script>
