+++ src/modules/home/what-are-doing/what-are-doing.astro
@@ -1,9 +1,9 @@
 ---
+import PointsRight from "@components/background/points-right.astro";
+import H2 from "@components/titles/H2.astro";
+import type { TLanguage } from "@i18n/config/interfaces/languaje.interface";
 import { translator } from "@i18n/config/translator";
 import Carrusel from "./carrusel.astro";
-import type { TLanguage } from "@i18n/config/interfaces/languaje.interface";
-import H2 from "@components/titles/H2.astro";
-import PointsRight from "@components/background/points-right.astro";
 
 const lang = Astro.params.lang as TLanguage;
 const t = translator(lang);
