+++ src/styles/themes.css
@@ -7,17 +7,17 @@
 
 @layer base {
     html[data-theme="theme1"] {
-        --color-primary: #00F5A0;
-        --color-secondary: #00D9FF;
-        --color-dark: #111111;
-        --color-light: #FFFFFF;
-    }
-
-    html[data-theme="theme2"] {
         --color-primary: #0066FF;
         --color-secondary: #00FFC6;
         --color-dark: #0A0A0A;
         --color-light: #F8F8F8;
+    }
+
+    html[data-theme="theme2"] {
+        --color-primary: #00F5A0;
+        --color-secondary: #00D9FF;
+        --color-dark: #111111;
+        --color-light: #FFFFFF;
     }
 
     html[data-theme="theme3"] {
