+++ src/styles/utilities.css
@@ -7,4 +7,28 @@
         scrollbar-width: none;
         -ms-overflow-style: none;   
     }
+
+    .sparkle {
+        position: absolute;
+        width: 6px;
+        height: 6px;
+        background: white;
+        border-radius: 9999px;
+        opacity: 0;
+        box-shadow: 0 0 6px white;
+        animation: sparkleAnim 2s ease-in-out infinite;
+    }
+
+    @keyframes sparkleAnim {
+        0%,
+        100% {
+            opacity: 0;
+            transform: scale(0.5);
+        }
+
+        50% {
+            opacity: 0.8;
+            transform: scale(1.5);
+        }   
+    }
 }
